# docs/openapi/climate-api-v1.json
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "Climate Monitoring System API",
                 "description":  "Public API exposed by Climate.Gateway. Internal service endpoints are intentionally excluded.",
                 "version":  "v1"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Local Docker Gateway"
                    }
                ],
    "paths":  {
...
                  "/api/alerts":  {
                                      "get":  {
                                                  "tags":  [
                                                               "Alerts"
                                                           ],
                                                  "parameters":  [
                                                                     {
                                                                         "name":  "riskType",
                                                                         "in":  "query",
                                                                         "schema":  {
                                                                                        "$ref":  "#/components/schemas/RiskType"
                                                                                    }
                                                                     },
                                                                     {
                                                                         "name":  "alertLevel",
                                                                         "in":  "query",
                                                                         "schema":  {
                                                                                        "$ref":  "#/components/schemas/AlertLevel"
                                                                                    }
                                                                     },
                                                                     {
                                                                         "name":  "sensorId",
                                                                         "in":  "query",
                                                                         "schema":  {
                                                                                        "type":  "string",
                                                                                        "format":  "uuid"
                                                                                    }
                                                                     },
                                                                     {
                                                                         "name":  "communityId",
                                                                         "in":  "query",
                                                                         "schema":  {
                                                                                        "type":  "string",
                                                                                        "format":  "uuid"
                                                                                    }
                                                                     },
                                                                     {
                                                                         "name":  "isActive",
                                                                         "in":  "query",
                                                                         "schema":  {
                                                                                        "type":  "boolean"
                                                                                    }
                                                                     }
                                                                 ],
                                                  "responses":  {
                                                                    "200":  {
                                                                                "description":  "OK",
                                                                                "content":  {
                                                                                                "text/plain":  {
                                                                                                                   "schema":  {
                                                                                                                                  "type":  "array",
                                                                                                                                  "items":  {
                                                                                                                                                "$ref":  "#/components/schemas/AlertResponse"
                                                                                                                                            }
                                                                                                                              }
                                                                                                               },
                                                                                                "application/json":  {
                                                                                                                         "schema":  {
                                                                                                                                        "type":  "array",
                                                                                                                                        "items":  {
                                                                                                                                                      "$ref":  "#/components/schemas/AlertResponse"
                                                                                                                                                  }
                                                                                                                                    }
                                                                                                                     },
                                                                                                "text/json":  {
                                                                                                                  "schema":  {
                                                                                                                                 "type":  "array",
                                                                                                                                 "items":  {
                                                                                                                                               "$ref":  "#/components/schemas/AlertResponse"
                                                                                                                                           }
                                                                                                                             }
                                                                                                              }
                                                                                            }
                                                                            },
                                                                    "401":  {
                                                                                "description":  "Unauthorized"
                                                                            },
                                                                    "403":  {
                                                                                "description":  "Forbidden"
                                                                            }
                                                                },
                                                  "operationId":  "Alerts_get_api_alerts",
                                                  "security":  [
                                                                   {
                                                                       "Bearer":  [

                                                                                  ]
                                                                   }
                                                               ]
                                              }
                                  },
                  "/api/alerts/{id}":  {
                                           "get":  {
                                                       "tags":  [
                                                                    "Alerts"
                                                                ],
                                                       "parameters":  [
                                                                          {
                                                                              "name":  "id",
                                                                              "in":  "path",
                                                                              "required":  true,
                                                                              "schema":  {
                                                                                             "type":  "string",
                                                                                             "format":  "uuid"
                                                                                         }
                                                                          }
                                                                      ],
                                                       "responses":  {
                                                                         "200":  {
                                                                                     "description":  "OK",
                                                                                     "content":  {
                                                                                                     "text/plain":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/AlertResponse"
                                                                                                                                   }
                                                                                                                    },
                                                                                                     "application/json":  {
                                                                                                                              "schema":  {
                                                                                                                                             "$ref":  "#/components/schemas/AlertResponse"
                                                                                                                                         }
                                                                                                                          },
                                                                                                     "text/json":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/AlertResponse"
                                                                                                                                  }
                                                                                                                   }
                                                                                                 }
                                                                                 },
                                                                         "401":  {
                                                                                     "description":  "Unauthorized"
                                                                                 },
                                                                         "403":  {
                                                                                     "description":  "Forbidden"
                                                                                 }
                                                                     },
                                                       "operationId":  "Alerts_get_api_alerts_id",
                                                       "security":  [
                                                                        {
                                                                            "Bearer":  [

                                                                                       ]
                                                                        }
                                                                    ]
                                                   }
                                       },
                  "/api/alerts/{id}/resolve":  {
                                                   "patch":  {
                                                                 "tags":  [
                                                                              "Alerts"
                                                                          ],
                                                                 "parameters":  [
                                                                                    {
                                                                                        "name":  "id",
                                                                                        "in":  "path",
                                                                                        "required":  true,
                                                                                        "schema":  {
                                                                                                       "type":  "string",
                                                                                                       "format":  "uuid"
                                                                                                   }
                                                                                    }
                                                                                ],
                                                                 "responses":  {
                                                                                   "200":  {
                                                                                               "description":  "OK"
                                                                                           },
                                                                                   "401":  {
                                                                                               "description":  "Unauthorized"
                                                                                           },
                                                                                   "403":  {
                                                                                               "description":  "Forbidden"
                                                                                           }
                                                                               },
                                                                 "operationId":  "Alerts_patch_api_alerts_id_resolve",
                                                                 "security":  [
                                                                                  {
                                                                                      "Bearer":  [

                                                                                                 ]
                                                                                  }
                                                                              ]
                                                             }
                                               },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "AlertLevel":  {
                                                          "enum":  [
                                                                       "Green",
                                                                       "Yellow",
                                                                       "Orange",
                                                                       "Red"
                                                                   ],
                                                          "type":  "string"
                                                      },
                                       "AlertResponse":  {
                                                             "type":  "object",
                                                             "properties":  {
                                                                                "id":  {
                                                                                           "type":  "string",
                                                                                           "format":  "uuid"
                                                                                       },
                                                                                "sensorId":  {
                                                                                                 "type":  "string",
                                                                                                 "format":  "uuid"
                                                                                             },
                                                                                "communityId":  {
                                                                                                    "type":  "string",
                                                                                                    "format":  "uuid"
                                                                                                },
                                                                                "alertType":  {
                                                                                                  "$ref":  "#/components/schemas/RiskType"
                                                                                              },
                                                                                "level":  {
                                                                                              "$ref":  "#/components/schemas/AlertLevel"
                                                                                          },
                                                                                "title":  {
                                                                                              "type":  "string",
                                                                                              "nullable":  true
                                                                                          },
                                                                                "description":  {
                                                                                                    "type":  "string",
                                                                                                    "nullable":  true
                                                                                                },
                                                                                "sensorValue":  {
                                                                                                    "type":  "number",
                                                                                                    "format":  "double"
                                                                                                },
                                                                                "thresholdValue":  {
                                                                                                       "type":  "number",
                                                                                                       "format":  "double"
                                                                                                   },
                                                                                "generatedAt":  {
                                                                                                    "type":  "string",
                                                                                                    "format":  "date-time"
                                                                                                },
                                                                                "isActive":  {
                                                                                                 "type":  "boolean"
                                                                                             },
                                                                                "resolvedAt":  {
                                                                                                   "type":  "string",
                                                                                                   "format":  "date-time",
                                                                                                   "nullable":  true
                                                                                               }
                                                                            },
                                                             "additionalProperties":  false
                                                         },
                                       "RiskType":  {
                                                        "enum":  [
                                                                     "Flood",
                                                                     "Drought",
                                                                     "Storm",
                                                                     "Frost",
                                                                     "ForestFire"
                                                                 ],
                                                        "type":  "string"
                                                    },
...
                                   },
                       "securitySchemes":  {
                                               "Bearer":  {
                                                              "type":  "http",
                                                              "scheme":  "bearer",
                                                              "bearerFormat":  "JWT",
                                                              "description":  "Paste only the accessToken returned by POST /api/auth/login. Do not include the 'Bearer ' prefix; Swagger UI adds it automatically."
                                                          }
                                           }
                   },
...
}